refactored to clean up components and fixed keys
